% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/title_or_id.R
\name{find_by_id}
\alias{find_by_id}
\title{Retrieve OMDB info by IMDB ID search}
\usage{
find_by_id(id, type = NULL, year_of_release = NULL, plot = "short",
  include_tomatoes = FALSE)
}
\arguments{
\item{id}{A valid IMDb ID (e.g. \code{tt1285016})}

\item{type}{Type of result to return. One of \code{[movie|series|episode|game]}.}

\item{year_of_release}{Year of release.}

\item{plot}{Return \code{short} or \code{full} plot.}

\item{include_tomatoes}{Include Rotten Tomatoes ratings.}
}
\value{
\code{tbl_df}/\code{data.frame} of search results (an empty one if none found). Also
        classes as an \code{omdb} object for prettier printing
}
\description{
Retrieve OMDB info by IMDB ID search
}
\note{
The API's search results limit to 10.
}
\seealso{
\href{omdbAPI documentation}{http://www.omdbapi.com/}
}

