% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gets.R
\name{get_countries}
\alias{get_countries}
\title{Get countries from an omdb object as a vector}
\usage{
get_countries(omdb)
}
\arguments{
\item{omdb}{an object returned by \code{find_by_title} or \code{find_by_id}}
}
\value{
A character vector of countries
}
\description{
Splits the field at commas and returns a vector of individual entries
}
\examples{
get_countries(find_by_id("tt0031381"))
}
