% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_imdb.R
\name{get_tv_data}
\alias{get_tv_data}
\title{Retrieve IMDB ratings for all Episodes of a TV series}
\usage{
get_tv_data(name)
}
\arguments{
\item{name}{TV series for which to obtain ratings for. Has to be of IMDB type "TV series".}
}
\value{
\code{data.frame}
}
\description{
Retrieve IMDB ratings for all Episodes of a TV series
}
\note{
Can be quite slow, depending on your internet connection.
}
\seealso{
\href{omdbAPI documentation}{http://www.omdbapi.com/}
}

