% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gets.R
\name{get_writers}
\alias{get_writers}
\title{Get writers from an omdb object as a vector}
\usage{
get_writers(omdb)
}
\arguments{
\item{omdb}{an object returned by \code{find_by_title} or \code{find_by_id}}
}
\value{
A character vector of writers
}
\description{
Splits the field at commas and returns a vector of individual entries
}
\examples{
get_writers(find_by_id("tt0031381"))
}
