% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/omdbapi-package.r
\docType{package}
\name{omdbapi}
\alias{omdbapi}
\alias{omdbapi-package}
\title{A package to access movie, television and game metadat via the Open Movie Database API}
\description{
A package to access movie, television and game metadat via the Open Movie Database API
}
\author{
Bob Rudis (@hrbrmstr)
}

