% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/title_or_id.R
\name{print.omdb}
\alias{print.omdb}
\title{Print an omdb result}
\usage{
\method{print}{omdb}(x, ...)
}
\arguments{
\item{x}{omdb object}

\item{\dots}{ignored}
}
\description{
If either \code{find_} function finds an omdb entry, the API only returns
a single record. That makes it possible to make a nice \code{print} routine
for it so the output is easier to read in interactive mode.
}

