% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/title_or_id.R
\name{search_by_title}
\alias{search_by_title}
\title{Lightweight omdb title search}
\usage{
search_by_title(term, type = NULL, year_of_release = NULL, page = 1)
}
\arguments{
\item{term}{Movie title to search for.}

\item{type}{Type of result to return. One of \code{[movie|series|episode|game]}.}

\item{year_of_release}{Year of release.}

\item{page}{1 for first 10 results, 2 for next ten, etc...}
}
\value{
\code{tbl_df}/\code{data.frame} of search results (an empty one if none found)
}
\description{
Lightweight omdb title search
}
\note{
The API's search results limit to 10 at a time (see argument page).
}
\seealso{
\href{omdbAPI documentation}{http://www.omdbapi.com/}
}

